Taking care of code review comments from Erik:
1) created utils package and moved utility routines under it
    - verifierutils.go
    - appinstanceutils.go
2) Fixed checkDiskSize to not take AppInstanceStatus as an argument.
3) Changed some Infof statements to Debugf statements to reduce noise
   in logs.
